Fix whitespace issues in SQLite absolute paths in DATABASE_URL
When the DSN is in the form "sqlite:/absolute/path" or
"sqlite://absolute/path" or "sqlite:///absolute/path", url.Parse
will consider the file path as :
 - "absolute" as the hostname
 - "path" (and the rest until "?") as the URL path.
Instead, when the DSN is in the form "sqlite:", the (relative) file
path is stored in the "Opaque" field.
See: https://pkg.go.dev/net/url#URL

While Opaque is not escaped, the URL Path is. So, in absolute paths
the full file path is being escaped as if it was an URL path.
This commit works around this behavior in ConnectionString (sqlite.go)
so that the return string is always a correct (url-unescaped) file path.