Fix error message assertion in SQLite test
This commit fixes the error message text assertion for DumpSchema (SQLite)
errors.
The DumpSchema returns an error with the connection string (aka the file
path) when there is an error while opening the database. Specifically, in
this case, the DumpSchema was tested to check whether an invalid file
was triggering the error.

The test is failing because, due to the previous commit, the connection
string is now rebuilt when url.Parse detects an URL path. While other
cases are covered by unit tests, the case of "." as file path was not
tested (no need to, it's not valid). In this case, the ConnectionString
function returns "/." as a connection string instead of ".".

While the ConnectionString behavior is wrong with ".", this fix doesn't
address that, as the connection string will be invalid anyway. So, I fixed
the error message instead.