Spanner Postgres interface support (#555)
Adds support for the [GCP Spanner Postgres
Interface](https://cloud.google.com/spanner/docs/postgresql-interface),
which only requires a slight modification to the postgres driver.

This will not work with a GoogleSQL-flavored Spanner database; support
for that will require a dedicated driver.

The new functionality is accessed with a custom `spanner-postgres://`
scheme. This is rather un-standard and something I just made up, so the
exact scheme is open for debate. We shouldn't use just `spanner://`,
since that may cause confusion down the line when a driver for Spanner
with the GoogleSQL dialect is added.

Related discussion: https://github.com/amacneil/dbmate/discussions/369