Enable FTS5 in SQLite migrations (#582)
This statically compiles in the FTS5 module, allowing dbmate to handle
migrations that use the full-text search features of SQLite.

## History

Issue #237 raised the question of handling migrations that included FTS5
tables. It was later converted to a discussion (#357).

## Solution

It is possible to enable the feature in the statically-compiled SQLite
that ships with dbmate via a feature flag. Adding the `sqlite_fts5` flag
to the `Makefile` embeds this feature statically.

## Testing

An additional test was added with a `CREATE VIRTUAL TABLE` statement
that leverages the FTS5 module. This test fails when dbmate is compiled
_without_ the feature flag. After adding the flag, this test passes.

`make test` and `make docker-all` pass.

Confirmed that it works in practice locally on Ubuntu 22.04.  

## Additional info

For information about the extension:
https://www.sqlite.org/fts5.html

For the feature flag for mattn/go-sqlite3:

https://github.com/mattn/go-sqlite3/?tab=readme-ov-file#feature--extension-list