Set mysqldump SSL/TLS args based on 'tls' connection URL param.
It looks like the default for `--ssl-verify-server-cert` was changed
from `FALSE` to `TRUE` in MariaDB Connector/C in version 3.4,
corresponding to MariaDB 11.4:

https://github.com/mariadb-corporation/mariadb-connector-c/commit/1287c901dc8515823d28edcebfe4be65e6c5a6b3

> Since version 3.4 peer certificate verification is enabled by default.

https://mariadb.com/docs/server/security/securing-mariadb/securing-mariadb-encryption/data-in-transit-encryption/securing-connections-for-client-and-server#enabling-one-way-tls-for-mariadb-clients

> Starting from [MariaDB 11.4](https://mariadb.com/docs/release-notes/community-server/mariadb-11-4-series/what-is-mariadb-114)
> (Connector/C version 3.4) this mode is enabled by default.

As `dbmate` uses the `go-sql-driver/mysql` driver for executing queries,
which sets [tls=false](https://github.com/go-sql-driver/mysql?tab=readme-ov-file#tls)
by default, we don't see a change when applying migrations.

However, `dbmate` executes `mysqldump` to dump schemas, and that's where
the change in MariaDB hits us.

We should disable SSL/TLS when invoking `mysqldump` if `tls` is `false`,
and we should use `--ssl-verify-server-cert=false` if `tls` is
`skip-verify`.